PNG Test Corpus
This is an assortment of PNG types. It's meant to test compression
algorithms to see if they can detect that, although the image is stored
as an RGBA, it only has 16 colors, and no transparency (and similar
tests). If the algorithm can then convert it to the correct type of PNG,
it will get a savings bonus in size without any loss in quality. This
also includes the PNG's in the smallest state to allow testing to see if
FLIF has different outputs depending on the PNG type.